Move reducer to slice
Move reducer to slice.
